Renaming Substitute to InvocationHandler
The previous concept of a Substitute was really responsible for handling invocations. The substitute itself should actually be a combination of a dynamic proxy to conform to the shape of the type being substituted for, an invocation handler to perform recording, setting results and potentially running queries, and an interceptor to forward invocations from the dynamic proxy to the invocation handler.

- Renamed almost occurrence of "call" to "invocation".
- Renamed Substitute to InvocationHandler
- Created new Substitute class as static gateway for NSubstitute functionality.
- Updated tests to reflect the modified terminology.