Fix TemporaryChange syntax
Updated TemporaryChange syntax to use expression trees rather than the previous, clumsy approach of specifying initial value, new value, and an Action<T> to set the values.

This new way is not as flexible or robust, but reads better for the scenarios in which we are using it, and is a much easier syntax to use.
Some more error checking would be good but at this stage only using it for very basic stuff so shouldn't cause too many problems.
This basic approach is shamelessly ripped off from JP Boodhoo's developwithpassion.bdd library (http://github.com/developwithpassion/developwithpassion.bdd).