Implementing substituting for delegates
Added CallRouterResolver.
SubstituteFactory registers ICallRouters for substitutes with its ICallRouterResolver and can use this dependency to return the ICallRouter used for each substitute.
SubstitutionContext no longer aware of how to cast substitutes to ICallRouters. Instead it delegates to SubstitutionFactory.
Overly simplistic implementation of CallRouterResolver. It casts proxies to ICallRouter where possible, otherwise it stores mappings between proxies and ICallRouters in an internal dictionary. This means that any delegate substituted for will be kept alive forever, along with their ICallRouter. This can be fixed by using WeakReferences.