Excluding property getter calls from Received.InOrder queries
As shown in the new test in SequenceChecking fixture, if the code
under test pulls out an intermediary variable for repeated property
access, this will cause what looks like a reasonable Received.InOrder
query to fail due to the differences in calls to property getters.

This seems like an failure due to over-specification, so will exclude
getters for this type of query.