Fixed regression where Received() clears stubbed value
Addition of ReturnAutoValueForThisAndSubsequentCallsHandler to
CheckReceivedCalls route (to allow awaiting a Received call) meant
previously stubbed values were lost. Changed this handler to support
forgetting the autovalue as well as using the value for subsequent calls.

- Introduced AutoValueBehaviour to handler and renamed to ReturnAutoValue.
- Updated RouteFactory to UseValueForSubsequentCalls for all routes except
  CheckReceivedCalls route. That route will forget whatever value is
  returned.

Note: could also fix this by running ReturnConfiguredResultHandler first.
The problem with this approach is it would cause .Returns callbacks to
run in cases where they are not expected.