Rework delegate proxy generation
Previously delegates were generated using the expressions, which
proxied all arguments to ICallRouter and later handled the result.
I've found that there is a better way - we can use the standard
Castle intercepting feature. The idea is that our delegate proxy
could be a regular delegate over the Caslte proxy's method.

That allowed to significantly improve the performance, as
expression compilation is pretty expensive.

Workflow:
1. Generate an interface for the requested delegate type, so it
contains the "Invoke" method with the required signature.
2. Ask Castle Proxy to generate a proxy for that interface.
3. Create a delegate from the "Invoke" method of the generated proxy.

It appears that this change is almost transparent for the engine
and works perfectly.