Rework SubstitutionContext to decrease coupling with the static members
Introduce the ThreadLocalContext class to hold the state bound
to the current thread. That breaks the circular dependency between
the SubstitutionContext and SubstituteFactory.

After the refactoring the static SubstitutionContext holds the
following members only:
- ThreadLocalContext to track thread-specific stuff
- SubstituteFactory to activate new substitutes
- CallRouterResolver to obtain router from the proxy

All the rest members became obsolete and refer to the inner ones.

This change allows to much simpler change the SubstituteFactory as
now it doesn't depend on the current SubstitutionContext.