Use statically defined IsReadOnlyAttribute type
Previously we generated the IsReadOnlyAttribute at run time.
The problem is that currenlty Castle.Core doesn't expose a way to synchronize
type creation. As result it could happen that type is defined twice and creation
fails with exception.

Given that we don't have a reliable way to ensure that type is defined once only,
we defined our own internal IsReadOnlyAttribute type and made it visible to the
dynamic assembly.

The reason this type is internal rather than public is that it might exist in the
target platform, so we don't want our attribute to interfere with system one.