Reorganize ArgumentSpecification factories to simplify tree
It appeared that actual tree of different factories
doesn't add to the code re-use, but significantly affects the code readability.
I merged most of the factories, to keep only a few ones, so it's much
easier to understand the code now. Also all the work related to argument
specifications dequeue and matching is now located in a single place.