Add protected `BaseResult()` method to CallInfo.
Create CallInfo<T> to calls that return results and expose `BaseResult`.
This gets messy to push the generic all the way through the code, so am
just using a cast in `Returns` extensions to handle this. This should be
safe as if we are in `Returns<T>` then the return value should be safe to
cast to a `T`.

Based off discussion here:
https://github.com/nsubstitute/NSubstitute/issues/622#issuecomment-640166693