Resolved a bug that would cause a struct to not match and throw a runtime error.
When a struct had an empty constructor that modified a field/property to a non-default value. The matcher would fail and throw a runtime error. This bug fix attempts to fix that issue by essentially zero initializing the struct the same way that default(T) does through the use of a runtime helper method. Simply, we replace Activator.CreateInstance with GetUninitializedObject.

The idea came from this post
https://stackoverflow.com/questions/1005336/c-sharp-using-reflection-to-create-a-struct

Which contains and deprecated method as of .NET8. Microsoft has a recommended fix for that here.
https://learn.microsoft.com/en-us/dotnet/fundamentals/syslib-diagnostics/syslib0050