Fixed bug that would cause a struct to not match and throw a runtime error. (#895)
* Resolved a bug that would cause a struct to not match and throw a runtime error.

When a struct had an empty constructor that modified a field/property to a non-default value. The matcher would fail and throw a runtime error. This bug fix attempts to fix that issue by essentially zero initializing the struct the same way that default(T) does through the use of a runtime helper method. Simply, we replace Activator.CreateInstance with GetUninitializedObject.

The idea came from this post
https://stackoverflow.com/questions/1005336/c-sharp-using-reflection-to-create-a-struct

Which contains and deprecated method as of .NET8. Microsoft has a recommended fix for that here.
https://learn.microsoft.com/en-us/dotnet/fundamentals/syslib-diagnostics/syslib0050

* Add tests for struct with default constructor

- Added test method `Any_on_struct_with_default_constructor_should_work` to ensure method calls with struct arguments do not throw exceptions.
- Introduced `StructWithDefaultConstructor` struct:
  - Contains a property `Value`.
  - Has a default constructor initializing `Value` to 42.
- Created `IWithStructWithDefaultConstructor` interface:
  - Declares `MethodWithStruct` that accepts `StructWithDefaultConstructor` as an argument.

* Fixed class constructor curly brace format

* Refactor default value handling logic

Updated `GetDefault` methods in `AutoObservableProvider.cs` and `AutoTaskProvider.cs` to use the `DefaultForType` class for determining default values, improving code modularity and reusability.